feat(minimal-0.1): implement P0 freeze — gold suite, std thin honesty
P0.1: fix std.core.ptr as ptrOffset shim; pointer remains canonical
P0.2: examples/minimal/* gold suite + cli_minimal_gold tests
P0.3: async await + multi-file spawn/join examples
OUT: experimental banners on env/fs/io/process/time/runtime extras
path: host-backed is_absolute (i64 ABI); stubs documented
freeze doc D2–D6 unblocked for installer phase
README already links freeze doc

+++ crates/arandu_backend_cranelift/src/rt_runtime.rs
@@ -119,12 +119,12 @@ pub unsafe extern "C" fn ar_rt_cancel_i64(handle: i64) {
 /// # Safety
 /// `ptr`/`len` fat string from Arandu JIT.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn ar_path_is_absolute(ptr: *const u8, len: i64) -> i32 {
+pub unsafe extern "C" fn ar_path_is_absolute(ptr: *const u8, len: i64) -> i64 {
     if len <= 0 || ptr.is_null() {
         return 0;
     }
     let s = unsafe { std::slice::from_raw_parts(ptr, len as usize) };
-    i32::from(s[0] == b'/')
+    i64::from(s[0] == b'/')
 }
 
 /// Path empty check.
@@ -132,8 +132,8 @@ pub unsafe extern "C" fn ar_path_is_absolute(ptr: *const u8, len: i64) -> i32 {
 /// # Safety
 /// Fat string ABI.
 #[unsafe(no_mangle)]
-pub unsafe extern "C" fn ar_path_is_empty(_ptr: *const u8, len: i64) -> i32 {
-    i32::from(len <= 0)
+pub unsafe extern "C" fn ar_path_is_empty(_ptr: *const u8, len: i64) -> i64 {
+    i64::from(len <= 0)
 }
 
 #[cfg(test)]
